pyproject.toml
[project]
name = "sherlock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.58.2",
    "chromadb>=1.0.15",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "google-cloud-storage>=3.2.0",
    "ipykernel>=6.29.5",
    "pydantic>=2.11.7",
    "pyperclip>=1.9.0",
    "python-slugify>=7.0.0",
]
